IMPALA-14594: Scale grouping aggregator's input cardinality by the number of backends
The grouping aggregator uses input cardinality in its decision
about whether to continue expanding for a pre-aggregation.
Currently, each fragment instance is using the full input
cardinality for this calculation, but it will only be processing
a fraction of the input. This makes pre-aggregations very aggressive
about expanding, because they assume very high levels of future
reduction.

This fixes the input cardinality calculation to scale it down
by the number of fragment instances for this fragment. Each
fragment instance now assumes it is processing 1/Nth of the
input cardinality (where N is the number of fragment instances).
This will make the pre-aggregation's expansion calculation
more accurate.

Testing:
 - Ran a core job

Change-Id: Iaf66709f12ea9b7328ca4e4c5cd45a05c9dc3032
Reviewed-on: http://gerrit.cloudera.org:8080/23732
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>