IMPALA-14564: Remove redundant partition info from Iceberg file descriptors
Iceberg file descriptors used to contain information about the
partition they belong to: the spec id and the partition values.
These fields uniquely identify the partition the file belongs
to and are only dependent on the partition not the file itself.
It means that it is redundant to store these fields in each
file descriptor in the Catalog.

This change normalizes partition-related metadata out of the
file descriptors in the Catalog and the Frontend.
The partition information is stored separately in the
IcebergContentFileStore in FlatBuffers binary format. File
descriptors only store a unique id of the partition they belong to.

Since the scan nodes need the partition information for execution,
we denormalize the file and partition metadata: we look up the
relevant partitions by their ids in the Planner, put them in the
file descriptors that are sent to the executors.
Metadata handling during scheduling and execution remains the same.

This change introduces a separate FlatBuffer schema to serialize
necessary Iceberg file metadata from the Catalog to the Frontend
and from the Frontend to the executors (in scan ranges) containing
only the relevant fields for maximum memory efficiency.

Testing:
 - ran existing e2e tests
 - extended FileMetadataLoaderTest to check the partitions

Assisted by Copilot (Claude Sonnet 4.6)
Change-Id: I57c2fd6f1ebb636aa9e7ca925413ca51858cbc2a
Reviewed-on: http://gerrit.cloudera.org:8080/23838
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>