IMPALA-14606: (addendum) Install Python 3 for RHEL8
The first IMPALA-14606 commit miss to setup Python 3 in fresh RHEL8
machine. This was not caught before because I test using downstream
jenkins and it reuse RHEL8 machine that previously setup with Python 2.

This patch fix the issue by skipping pip install argparse that broke the
script and run setup_python3 instead for RHEL8 machine.

Testing:
- Run full bootstrap_system.sh and buildall.sh in fresh RHEL8 machine.

Change-Id: I6df0a534175404fe96d32eeb1e7bf0aa9ca204cd
Reviewed-on: http://gerrit.cloudera.org:8080/23772
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Laszlo Gaal <laszlo.gaal@cloudera.com>
Tested-by: Riza Suminto <riza.suminto@cloudera.com>