IMPALA-14628: Fix mismatch between greatest synced event time and last synced event time
The greatest synced event id and its corresponding event time
must match the last synced event id and time once all outstanding
events have been fully processed.

Previously, for batch events, timestamp of the first event in
the batch is recorded when transitioning the batch event from
the in-progress log to the processed log in EventExecutorService.
As a result, the greatest synced event time incorrectly reflected
the first event’s timestamp, even though the greatest synced
event id corresponded to the last event in the batch, leading to
a mismatch.

This fix ensures that, for batch events, the timestamp of the
last event in the batch is recorded when the batch is marked as
processed. Consequently, the greatest synced event time now
correctly aligns with the last synced event id after all
outstanding events are processed.

Testing:
    - Added an FE test and ran the existing tests.

Change-Id: I26550441bc0d5e7736c3b900c3e5178103157446
Reviewed-on: http://gerrit.cloudera.org:8080/23797
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>