IMPALA-14647: Fix truncate for replicated txn tables always delete data
If a transactional table is being replicated by Hive replication, Impala
uses the truncateTable HMS API to truncate it. However, the last
parameter (boolean deleteData) is not set, which causing HMS always
delete the data.

In HiveServer, there is a configuration, "hive.acid.truncate.usebase",
to control whether to create a new base_* dir (so keep the data files)
for truncate operation. This patch fixes the API invocation to respect
this configuration.

Note that for non-transactional tables, the behavior is still deleting
the data files, which is consistent with Hive.

Tests:
 - Added test to verify the data files exist after truncate.
 - Added util method to fetch effective hadoop and hive configs from the
   /hadoop-varz endpoint.

Change-Id: Ia31991baeb2ef8717c387b841b65cff562dbcae0
Reviewed-on: http://gerrit.cloudera.org:8080/23810
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>