IMPALA-14646: StorageDescriptor normalization should deal with parameters
When checking whether an ALTER_TABLE event has trivial changes in
StorageDescriptor, we normalize fields that are unrelated to file
metadata, e.g. cols, bucketCols, sortCols, etc. However, the parameters
map of StorageDescriptor is not normalized, which causes null and empty
map be considered as different.

This patch adds the normalization on the parameters map of
StorageDescriptor. Also improves the logs when non-trival SD changes is
detected. Currently we just dump the full SD objects, which is pretty
verbose and hard to analyze. This patches add logs to show the actual
changes.

Tests
 - Added FE test

Change-Id: I6a9fcf2d60a41e9669d49412d49a2416c13d17bc
Reviewed-on: http://gerrit.cloudera.org:8080/23814
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>