IMPALA-13679: Fix ASAN Failure in Workload Management Tests
The TickerSecondsBool class defined in ticker.h is used to
process queued completed queries at specified intervals.
This class references variables defined outside of it. The
workload management code never properly shuts down this
ticker resulting in it sometimes, during shutdown, to
access variables that have already been released.

Removes the Ticker class leveraging the available
primitives of the std::condition_variable type.

The Ticker class was originally developed to avoid spurious
wakeups of a std::condition_variable, but the functionality
provided by the Ticker class can be eliminated by leveraging
the built in functionality of std::condition_variable.

In the workload-management-worker.cc file, adds the static
keyword to variables that are intended to be used only in
that file.

Testing
* workload management custom cluster tests passed locally
  and in jenkins build
* TSAN/ASAN builds passed with workload management enabled
  on the EE test impalad

Change-Id: I9a29f4f788cf7259c263a17272e86acdda5caa77
Reviewed-on: http://gerrit.cloudera.org:8080/23818
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>